test: enable CodeHint-integ-test
